Change tensor type conversion from float64 to float32

+++ src/data/generate_datasets.py
@@ -60,7 +60,7 @@ def main():
             elif isinstance(val, dict):
                 for subkey, subval in val.items():
                     if isinstance(subval, tuple): # Handling the bc dictionary structure
-                        case_data[key][subkey] = tuple(t.to(torch.float64) for t in subval)
+                        case_data[key][subkey] = tuple(t.to(torch.float32) for t in subval)
 
         # Calculate and store ic_true directly in the dictionary
         U0_val = case["U0"]
